creates farmer top up activity

+++ app/src/main/java/com/mobile/wanda/promoter/model/requests/VoucherTopUpRequest.kt
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName
  */
 data class VoucherTopUpRequest(
         @SerializedName("user_id")
-        val user_id: Int,
+        val user_id: Long,
         @SerializedName("payment_option")
         val paymentOption: String,
         @SerializedName("amount")
